moved primitive util dir
