+++ featuretools/computational_backends/pandas_backend.py
@@ -16,7 +16,7 @@ from .feature_tree import FeatureTree
 
 from featuretools import variable_types
 from featuretools.exceptions import UnknownFeature
-from featuretools.primitive_utils import (
+from featuretools.primitives.primitive_utils import (
     AggregationPrimitive,
     DirectFeature,
     IdentityFeature,
