+++ featuretools/primitives/api.py
@@ -1,11 +1,11 @@
 # flake8: noqa
-from .aggregation_primitives import *
-from .binary_transform import *
-from .cum_transform_feature import *
-from .installed import *
-from .transform_primitive import *
-from .utils import (
+from .builtin_primitives.aggregation_primitives import *
+from .builtin_primitives.binary_transform import *
+from .builtin_primitives.cum_transform_feature import *
+from .builtin_primitives.transform_primitive import *
+from .builtin_primitives.utils import (
     get_aggregation_primitives,
     get_transform_primitives,
     list_primitives
 )
+from .installed import *
