+++ featuretools/primitives/builtin_primitives/cum_transform_feature.py
@@ -6,14 +6,14 @@ import pandas as pd
 
 from .aggregation_primitives import Count, Max, Mean, Min, Sum
 
-from featuretools.primitive_utils.primitive_base import (
+from ..primitive_utils.primitive_base import (
     IdentityFeature,
     PrimitiveBase
 )
-from featuretools.primitive_utils.transform_primitive_base import (
+from ..primitive_utils.transform_primitive_base import (
     TransformPrimitive
 )
-from featuretools.primitive_utils.utils import apply_dual_op_from_feat
+from ..primitive_utils.utils import apply_dual_op_from_feat
 from featuretools.utils import is_string
 from featuretools.utils.wrangle import _check_timedelta
 from featuretools.variable_types import Id, Index, Numeric, TimeIndex
