+++ featuretools/primitives/installed/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
-from featuretools.primitive_utils import load_primitive_from_file, list_primitive_files, get_installation_dir, Feature
+from ..primitive_utils import load_primitive_from_file, list_primitive_files, get_installation_dir, Feature
 
 
 # iterate over files in installed, import class that are right subclass
