+++ featuretools/tests/primitive_tests/bad_primitive_files/multiple_primitives.py
@@ -1,4 +1,4 @@
-from featuretools.primitive_utils import make_agg_primitive
+from featuretools.primitives import make_agg_primitive
 from featuretools.variable_types import Numeric
 
 CustomMax = make_agg_primitive(lambda x: max(x),
